Don't invalidate judge data for judge being added to a second event

+++ app/technovation/invalidate_existing_judge_data.rb
@@ -1,7 +1,9 @@
 class InvalidateExistingJudgeData
   def self.call(participant)
-    participant.submission_scores.current_round.destroy_all
-    participant.judge_assignments.destroy_all
-    participant.events.destroy_all
+    if participant.is_team? or participant.virtual_event?
+      participant.submission_scores.current_round.destroy_all
+      participant.judge_assignments.destroy_all
+      participant.events.destroy_all
+    end
   end
 end
